Explicitly use size_type, no relying on implicit conversion from int.

+++ src/random.h
@@ -9,20 +9,20 @@
 #ifndef NEUROFIELD_SRC_RANDOM_H
 #define NEUROFIELD_SRC_RANDOM_H
 
-#include<random>
+#include <random>
 
 class Random {
   Random(Random& ); // no copy constructor
 
-  static int id; // Auto-increment the ID of each Random instance to use as the default seed
+  static uint_fast64_t id; // Auto-increment the ID of each Random instance to use as the default seed
 
   std::mt19937_64 gen; ///< Generator
   std::normal_distribution<double> dist; ///< Distribution
 
  public:
   Random(double mean, double std);
-  explicit Random(int seed, double mean = 0.0, double std = 1.0);
-  void init(int seed, double mean, double std);
+  explicit Random(uint_fast64_t seed, double mean = 0.0, double std = 1.0);
+  void init(uint_fast64_t seed, double mean, double std);
   double get(void);
   void get(double& target);
   ~Random(void);
